+++ lib/src/core/core.dart
@@ -1,8 +1,0 @@
-export 'dispatch.dart';
-export 'equality.dart';
-export 'internal/internal.dart';
-export 'loadable.dart';
-export 'middleware.dart';
-
-export 'commands/commands.dart';
-export 'store/store.dart';