+++ lib/src/flutter/flutter.dart
@@ -1,5 +1,3 @@
-export 'binding_utils.dart';
-export 'dispatch_proxy.dart';
 export 'helm_builder.dart';
 export 'helm_computed.dart';
 export 'helm_controller.dart';